Update HTTP status code messages for DELETE
When a successful delete occurs and we don't
send anything in response, the status code should
be 204. When the item that we want to delete cannot
be found, the status code should be 404. Also, we need
a response on the endpoint or else we get a gross error
in the Phoenix logs (Plug.Conn.NotSentError)